Better namespace structure for unit tests
